🐛   move sample_entropy function from stats.py to forecastability.py and update its implementation
